Added additional method in Date Helper class mockups are updated for singleton Added constructor for InputEntry to facilitate easy object creation Created a new QueueEvent model Change the return type of member variable in OutputPoint model SimulatorClient: updated for using new constructor for Input points SimulatorEngine : Changed logic to use Queue for processing on Input data points

+++ src/org/rohit/test/bloodsugar/model/OutputPoint.java
@@ -9,12 +9,12 @@ import java.util.Date;
  */
 public class OutputPoint {
 
-	private long value;
+	private Double value;
 	private Date timestamp;
-	public long getValue() {
+	public Double getValue() {
 		return value;
 	}
-	public void setBloodSugar(long value) {
+	public void setBloodSugar(double value) {
 		this.value = value;
 	}
 	public Date getTimestamp() {
